Added: New class SIMinput sitting between SIMbase and SIMoutput in the SIM-hierarchy. It contains all general model definition methods (input- file parsing, property setup, initial conditions, adaptive refinement).
Changed: Renamed the virtual createModelGenerator to getModelGenerator
plus some cosmetics in console feedback during default model generation.

The latter will require updates of the apps relying on model generators
and regression tests because of cosmetic correction of console feedback.
