+++ src/SIM/SIM2D.C
@@ -12,19 +12,17 @@
 //==============================================================================
 
 #include "SIM2D.h"
-#include "IFEM.h"
+#include "ModelGenerator.h"
 #include "ASMs2DC1.h"
 #include "ImmersedBoundaries.h"
 #include "Functions.h"
-#include "ModelGenerator.h"
 #include "Utilities.h"
 #include "Vec3Oper.h"
+#include "IFEM.h"
 #include "tinyxml.h"
 #ifdef USE_OPENMP
 #include <omp.h>
 #endif
-#include <fstream>
-#include <sstream>
 
 
 /*!
@@ -156,9 +154,10 @@ bool SIM2D::parseGeometryTag (const TiXmlElement* elem)
       for (int j = lowpatch; j <= uppatch; j++)
         if ((pch = dynamic_cast<ASM2D*>(this->getPatch(j,true))))
         {
-          IFEM::cout <<"\tRefining P"<< j <<" dir="<< dir;
+          IFEM::cout <<"\tRefining P"<< j <<" dir="<< dir
+                     <<" with grading "<< elem->FirstChild()->Value() <<":";
           for (size_t i = 0; i < xi.size(); i++)
-            IFEM::cout <<" "<< xi[i];
+            IFEM::cout << (i%10 || xi.size() < 11 ? " " : "\n\t") << xi[i];
           IFEM::cout << std::endl;
           pch->refine(dir-1,xi);
         }
@@ -372,10 +371,8 @@ bool SIM2D::parse (const TiXmlElement* elem)
     else if (!strcasecmp(elem->Value(),"boundaryconditions"))
       result &= this->parseBCTag(child);
 
-  if (myGen && result) {
-    if (!this->createFEMmodel()) return false;
-    myGen->createTopology(*this);
-  }
+  if (myGen && result)
+    result = myGen->createTopology(*this);
 
   delete myGen;
   myGen = nullptr;
@@ -719,10 +716,11 @@ bool SIM2D::addConstraint (int patch, int lndx, int ldim, int dirs, int code,
 
 
 ASMbase* SIM2D::readPatch (std::istream& isp, int pchInd,
-                           const std::vector<unsigned char>& unf) const
+                           const CharVec& unf) const
 {
-  const std::vector<unsigned char>& uunf = unf.empty()?nf:unf;
-  ASMbase* pch = ASM2D::create(opt.discretization,nsd,uunf,uunf.size() > 1);
+  const CharVec& uunf = unf.empty() ? nf : unf;
+  bool isMixed = uunf.size() > 1 && uunf[1] > 0;
+  ASMbase* pch = ASM2D::create(opt.discretization,nsd,uunf,isMixed);
   if (pch)
   {
     if (!pch->read(isp))
@@ -745,8 +743,9 @@ bool SIM2D::readPatches (std::istream& isp, PatchVec& patches,
                          const char* whiteSpace) const
 {
   ASMbase* pch = nullptr;
+  bool isMixed = nf.size() > 1 && nf[1] > 0;
   for (int pchInd = 1; isp.good(); pchInd++)
-    if ((pch = ASM2D::create(opt.discretization,nsd,nf,nf.size() > 1 && nf[1] > 0)))
+    if ((pch = ASM2D::create(opt.discretization,nsd,nf,isMixed)))
     {
       if (!pch->read(isp))
       {
@@ -757,7 +756,8 @@ bool SIM2D::readPatches (std::istream& isp, PatchVec& patches,
         delete pch;
       else
       {
-        IFEM::cout << whiteSpace <<"Reading patch "<< pchInd << std::endl;
+        if (whiteSpace)
+          IFEM::cout << whiteSpace <<"Reading patch "<< pchInd << std::endl;
         pch->idx = patches.size();
         patches.push_back(pch);
         if (checkRHSys)
@@ -829,9 +829,8 @@ void SIM2D::clonePatches (const PatchVec& patches,
 }
 
 
-ModelGenerator* SIM2D::createModelGenerator(const TiXmlElement* geo) const
+ModelGenerator* SIM2D::getModelGenerator (const TiXmlElement* geo) const
 {
-  IFEM::cout <<"  Using default linear geometry basis on unit domain [0,1]^2";
   return new DefaultGeometry2D(geo);
 }
