+++ src/SIM/SIMdependency.C
@@ -156,17 +156,3 @@ bool SIMdependency::extractPatchDependencies (IntegrandBase* problem,
 
   return true;
 }
-
-
-bool SIMdependency::hasIC (const std::string& name) const
-{
-  InitialCondMap::const_iterator it;
-  std::vector<ICInfo>::const_iterator it2;
-
-  for (it = myICs.begin(); it != myICs.end(); ++it)
-    for (it2 = it->second.begin(); it2 != it->second.end(); ++it2)
-      if (it2->sim_field == name)
-        return true;
-
-  return false;
-}
