+++ src/SIM/SIMdependency.h
@@ -32,28 +32,6 @@ public:
   //! \brief Spline patch container
   typedef std::vector<ASMbase*> PatchVec;
 
-  //! \brief Struct holding information about an initial condition.
-  struct ICInfo
-  {
-    int sim_level;  //!< The time level for the field in the SIM class
-    int file_level; //!< The time level for the field in the file
-    int geo_level;  //!< The time level for the geometry in the file
-    char basis;     //!< The basis to inject field into (for mixed)
-    int component;  //!< Component for field (for functions)
-    std::string sim_field;  //!< The name of the field in the SIM class
-    std::string file_field; //!< The field name in the file or type of function
-    std::string function;   //!< Function if given in function form
-    //! \brief Default constructor.
-    ICInfo() : sim_level(0), file_level(0), geo_level(0) {}
-    //! \brief Constructor providing the field name.
-    ICInfo(const std::string& f) : sim_level(0), file_level(0), geo_level(0),
-                                   basis(1), component(-1),
-                                   sim_field(f), file_field(f) {}
-  };
-
-  //! \brief Initial condition container
-  typedef std::map< std::string,std::vector<ICInfo> > InitialCondMap;
-
 private:
   //! \brief Struct holding information about an inter-SIM dependency.
   struct Dependency
@@ -81,12 +59,6 @@ protected:
 public:
   //! \brief Empty destructor.
   virtual ~SIMdependency() {}
-
-  //! \brief Returns a const reference to the initial condition container.
-  const InitialCondMap& getICs() const { return myICs; }
-
-  //! \brief Checks whether a named initial condition is present.
-  virtual bool hasIC(const std::string& name) const;
 
   //! \brief Returns the number of spatial dimensions in the model.
   virtual size_t getNoSpaceDim() const = 0;
@@ -121,6 +93,8 @@ public:
   ASMbase* getDependentPatch(const std::string& name, int pindx) const;
   //! \brief Registers a named field with associated nodal vector in this SIM.
   void registerField(const std::string& name, const utl::vector<double>& vec);
+  //! \brief Checks whether a named initial condition is present.
+  virtual bool hasIC(const std::string&) const { return false; }
 
 private:
   //! \brief Returns an iterator pointing to a named dependency.
@@ -137,9 +111,6 @@ protected:
 private:
   FieldMap  myFields;  //!< The named fields of this SIM object
   DepVector depFields; //!< Other fields this SIM objecy depends on
-
-protected:
-  InitialCondMap myICs; //!< The initial conditions
 };
 
 #endif
